Update README.md
The need for predictive maintenance system for wall surface buildings has emerged and several works have been done around developing and improving maintenance for buildings through the use of Artificial Intelligence, preferably Machine and deep learning algorithms. To able to build a maintenance system, deflection and Cracks were the yardstick for evaluating the structural health of a building. Image classification was done on the cracks dataset and prediction of deflection was achieved using the numerical dataset and set at a threshold based on the maximum permissible deflection for a wall surface building.